forecasting/Chronos/requirements.txt
torch==2.2.1
numpy==1.26.2
pandas==2.1.3
chronos-forecasting==2.2.2
SQLAlchemy==2.0.49
mysql-connector-python==9.7.0
fastapi==0.136.1
uvicorn==0.34.0